luma-backend/requirements.txt
# Luma MVP - Backend Requirements
# Python 3.11+

# FastAPI Framework
fastapi==0.109.0
uvicorn[standard]==0.27.0
python-multipart==0.0.6

# Database
sqlalchemy==2.0.25
psycopg2-binary>=2.9.10
asyncpg==0.29.0

# Supabase
supabase==2.3.0

# Authentication
pyjwt==2.8.0
python-jose[cryptography]==3.3.0
passlib[bcrypt]==1.7.4

# OCR & Document Processing
pytesseract==0.3.10
pdf2image==1.17.0
Pillow==10.2.0
google-cloud-vision==3.6.0

# Data Processing
pandas==2.2.0
openpyxl==3.1.2
xlrd==2.0.1

# PDF Generation
reportlab==4.0.9
weasyprint==60.2

# Utilities
python-dotenv==1.0.1
pydantic==2.5.3
pydantic-settings==2.1.0

# Development
pytest==7.4.4
pytest-asyncio==0.23.3
httpx==0.26.0
black==24.1.1
flake8==7.0.0
